modelmapper update

+++ bolhinhos/src/main/java/com/franciscoimbra/bolhinhos/repository/ClientUserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClientUserRepository extends JpaRepository<ClientUser, Integer> {
+public interface ClientUserRepository extends JpaRepository<ClientUser, Long> {
 }
