+++ bolhinhos/src/main/java/com/franciscoimbra/bolhinhos/repository/CustomerOrderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerOrderRepository extends JpaRepository<CustomerOrder, Integer> {
+public interface CustomerOrderRepository extends JpaRepository<CustomerOrder, Long> {
 }
